requirements.txt
azure-ai-ml
azure-identity
azure-mgmt-cognitiveservices
azure-search
azure-search-documents
BeautifulSoup4
requests
openai
python-dotenv